% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{captureSpecials}
\alias{captureSpecials}
\title{Capture Warnings and Errors}
\usage{
captureSpecials(expr)
}
\arguments{
\item{expr}{expression to evaluate.}
}
\description{
Converts warnings and errors into in-band information.
This function is for internal use and is used within
the attached R session in the background.
}
\keyword{internal}
