% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlrintermbo-package}
\alias{mlrintermbo}
\alias{mlrintermbo-package}
\title{mlrintermbo: An 'mlrMBO' 'mlr3' Interface}
\description{
Model-based optimization for 'mlr3' through 'mlrMBO'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mb706/mlrintermbo}
  \item Report bugs at \url{https://github.com/mb706/mlrintermbo/issues}
}

}
\author{
\strong{Maintainer}: Martin Binder \email{developer.mb706@doublecaret.com}

}
