% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{OptimInstanceMultiCrit}
\alias{OptimInstanceMultiCrit}
\title{OptimInstanceMultiCrit Class}
\description{
\code{bbotk}'s \code{OptimInstanceMultiCrit} class.
Re-exported since \code{bbotk} will change the name.
}
