% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{OptimInstanceSingleCrit}
\alias{OptimInstanceSingleCrit}
\title{OptimInstanceSingleCrit Class}
\description{
\code{bbotk}'s \code{OptimInstanceSingleCrit} class.
Re-exported since \code{bbotk} will change the name.
}
