% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerInterMBO.R
\name{OptimizerInterMBO}
\alias{OptimizerInterMBO}
\alias{mlr_optimizers_intermbo}
\alias{TunerInterMBO}
\title{Tuner using mlrMBO}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3tuning:Tuner]{mlr3tuning::Tuner}.
}
\description{
Subclass for mlrMBO tuning.

mlrMBO must not be loaded directly into R when using mlr3, for various reasons.
TunerInterMBO takes care that this does not happen.
}
\examples{
library("paradox")
library("bbotk")

# silly example function: minimize x^2 for -1 < x < 1
domain <- ParamSet$new(list(ParamDbl$new("x", lower = -1, upper = 1)))
codomain <- ParamSet$new(list(ParamDbl$new("y", tags = "minimize")))
objective <- ObjectiveRFun$new(function(xs) list(y = xs$x^2), domain, codomain)

# initialize instance
instance <- OptimInstanceSingleCrit$new(objective, domain, trm("evals", n_evals = 6))

# use intermbo optimizer
optser <- opt("intermbo")

# optimizer has hyperparameters from mlrMBO
optser$param_set$values$final.method <- "best.predicted"

# optimization happens here.
optser$optimize(instance)

instance$result
}
