% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{TuningInstanceMultiCrit}
\alias{TuningInstanceMultiCrit}
\title{TuningInstanceMultiCrit Class}
\description{
\code{mlr3tuning}'s \code{TuningInstanceMultiCrit} class.
Re-exported since \code{mlr3tuning} will change the name.
}
