% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{TuningInstanceSingleCrit}
\alias{TuningInstanceSingleCrit}
\title{TuningInstanceSingleCrit Class}
\description{
\code{mlr3tuning}'s \code{TuningInstanceSingleCrit} class.
Re-exported since \code{mlr3tuning} will change the name.
}
