% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlrintermbo-package}
\alias{mlrintermbo}
\alias{mlrintermbo-package}
\title{mlrintermbo: Model-Based Optimization for 'mlr3' Through 'mlrMBO'}
\description{
The 'mlrMBO' package can ordinarily not be used for optimization within 'mlr3', because of incompatibilities of their respective class systems. 'mlrintermbo' offers a compatibility interface that provides 'mlrMBO' as an 'mlr3tuning' 'Tuner' object, for tuning of machine learning algorithms within 'mlr3', as well as a 'bbotk' 'Optimizer' object for optimization of general objective functions using the 'bbotk' black box optimization framework. The control parameters of 'mlrMBO' are faithfully reproduced as a 'paradox' 'ParamSet'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mb706/mlrintermbo}
  \item Report bugs at \url{https://github.com/mb706/mlrintermbo/issues}
}

}
\author{
\strong{Maintainer}: Martin Binder \email{developer.mb706@doublecaret.com}

}
